Organised files in Application project
