+++ Core/Application/CQRS/Experiments/Queries/GetExperimentDetail/ExperimentDetailVm.cs
