+++ Core/Application/CQRS/Experiments/Queries/GetExperimentDetail/FactorDto.cs
