+++ Core/Application/CQRS/Experiments/Queries/GetExperimentDetail/GetExperimentDetailQuery.cs
