+++ Core/Application/CQRS/Experiments/Queries/GetExperimentDetail/GetExperimentDetailQueryHandler.cs
