+++ Core/Application/CQRS/Experiments/Queries/GetExperimentDetail/LevelDto.cs
