+++ Core/Application/CQRS/Experiments/Queries/GetExperimentDetail/TreatmentDto.cs
