+++ Core/Application/CQRS/Experiments/Queries/GetExperimentList/ExperimentLookupDto.cs
