+++ Core/Application/Common/EventManager.cs
@@ -11,7 +11,13 @@ namespace Rems.Application
     {
         public delegate void ItemNotFoundHandler(object sender, ItemNotFoundArgs args);
         public static event ItemNotFoundHandler ItemNotFound;
-        public static void InvokeItemNotFound(object sender, ItemNotFoundArgs args) => ItemNotFound?.Invoke(sender, args);
+        public static void InvokeItemNotFound(object sender, ItemNotFoundArgs args) 
+            => ItemNotFound?.Invoke(sender, args);
+
+        public delegate void ProgressTrackingActivatedHandler(object sender, ProgressTrackingActivatedArgs args);
+        public static event ProgressTrackingActivatedHandler ProgressTrackingActivated;
+        public static void InvokeProgressTrackingActivated(object sender, ProgressTrackingActivatedArgs args) 
+            => ProgressTrackingActivated?.Invoke(sender, args);
     }
 
     public class ItemNotFoundArgs : EventArgs
@@ -23,5 +29,10 @@ namespace Rems.Application
         public string Selection { get; set; }
 
         public bool Cancelled { get; set; } = false;
+    }
+
+    public class ProgressTrackingActivatedArgs : EventArgs
+    {
+        public int Count { get; set; }
     }
 }
