+++ Core/Application/Common/Extensions/DataTableExtensions.cs
@@ -2,10 +2,11 @@
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data;
+using System.Linq;
 
 namespace Rems.Application.Common.Extensions
 {
-    static class DataTableExtensions
+    public static class DataTableExtensions
     {
         public static DataTable ToDataTable<T>(this IEnumerable<T> data)
         {
@@ -22,6 +23,49 @@ namespace Rems.Application.Common.Extensions
                 table.Rows.Add(row);
             }
             return table;
+        }
+
+        public static IEnumerable<DataRow> DistinctRows(this DataTable table)
+        {
+            var comparer = new DataRowItemComparer();
+            var rows = table.Rows.Cast<DataRow>();
+
+            return rows.Distinct(comparer);
+        }
+    }
+
+    public class DataRowItemComparer : IEqualityComparer<DataRow>
+    {
+        public bool Equals(DataRow x, DataRow y)
+        {
+            var xItems = x.ItemArray;
+            var yItems = y.ItemArray;
+
+            if (xItems.Length != yItems.Length) return false;
+
+            for (int i = 0; i < xItems.Length; i++)
+            {
+                if (!Equals(xItems[i], yItems[i])) return false;
+            }
+
+            return true;
+        }
+
+        public int GetHashCode(DataRow obj)
+        {
+            unchecked
+            {
+                if (obj.ItemArray == null) return 0;
+
+                int hash = 17;
+
+                foreach (var o in obj.ItemArray)
+                {
+                    hash *= 31;
+                    hash += o.GetHashCode();
+                }
+                return hash;
+            }
         }
     }
 }
